+++ src/Services/Sale/BasketItem/Result/DeletedBasketItemResult.php
@@ -1,0 +1,26 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Sally Fancen <vadimsallee@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Sale\BasketItem\Result;
+
+use Bitrix24\SDK\Core\Result\DeletedItemResult;
+
+/**
+ * Class DeletedBasketItemResult
+ * Represents the result of deleting a basket item
+ *
+ * @package Bitrix24\SDK\Services\Sale\BasketItem\Result
+ */
+class DeletedBasketItemResult extends DeletedItemResult
+{
+}